deleted spring cloud dependecies, changed PostSummary's model etc..

+++ phots-up-back-end/src/main/java/main/dao/post/PostDao.java
@@ -15,7 +15,7 @@ public interface PostDao {
 	
 	public Optional<Post> deletePost(Long postId);
 	
-	public int updatePost(Post post);
+	public long updatePost(Post post);
 	
 	public boolean addLike(Long postId, Long userId);
